// packages/rural-event-categories/src/types/googleNaturalLanguageCategories.ts
export type GoogleNaturalLanguageL1 =
  | "/Adult"
  | "/Arts & Entertainment"
  | "/Autos & Vehicles"
  | "/Beauty & Fitness"
  | "/Books & Literature"
  | "/Business & Industrial"
  | "/Computers & Electronics"
  | "/Finance"
  | "/Food & Drink"
  | "/Games"
  | "/Health"
  | "/Hobbies & Leisure"
  | "/Home & Garden"
  | "/Internet & Telecom"
  | "/Jobs & Education"
  | "/Law & Government"
  | "/News"
  | "/Online Communities"
  | "/People & Society"
  | "/Pets & Animals"
  | "/Real Estate"
  | "/Reference"
  | "/Science"
  | "/Sensitive Subjects"
  | "/Shopping"
  | "/Sports"
  | "/Travel";